Started work on messenger server
Wrote basic foundation that we can use, not sure if it works correctly yet because I ran out of time to check it, but its a place to start. Will edit client file later, but for now has a framework that I got from somewhere else that we can build upon.